Ensure shutdown run before nfs-server stops (desc)
With the introduction in openflight-slurm-multinode cluster type there
will be the possibility for anything to be an NFS server. In the event
that something sharing NFS is shut down then, to prevent other systems
from locking up, the remove process will need to do a umount on all
clients.

This change adds the nfs-server service as a dependency for shutdown,
this works even on systems that aren't NFS servers as it will simply be
ignored. (Tested by adding gibberish non-existent service to the
shutdown script and observing that it is ignored)